1. Updated to newest OllamaSharp & OpenAi 2. Reworked Ollama and OpenAi plugins

+++ Musoq.DataSources.OpenAI/OpenAiRequestInfo.cs
@@ -4,11 +4,11 @@ internal class OpenAiRequestInfo
 {
     public string Model { get; init; } = string.Empty;
     
-    public double FrequencyPenalty { get; init; } = 0.0;
+    public float FrequencyPenalty { get; init; } = 0;
     
     public int MaxTokens { get; init; } = 4000;
     
-    public double PresencePenalty { get; init; } = 0.0;
+    public float PresencePenalty { get; init; } = 0;
     
-    public double Temperature { get; init; } = 0.0;
+    public float Temperature { get; init; } = 0;
 }
